requirements.txt
certifi==2018.11.29
chardet==3.0.4
cx-Oracle==7.0.0
DBUtils==1.3
idna==2.8
numpy==1.16.1
pandas==0.24.1
psutil==5.5.0
pymongo==3.7.2
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.9
redis==3.1.0
requests==2.21.0
schedule==0.6.0
simplejson==3.16.0
six==1.12.0
urllib3==1.24.1
